refactor: add rsvp list to it's own component
